fix: don't prefix locale to 'mailto' links
We don't prefix the locale code to 'mailto' links in the header and the
footer.

This also updates the wording of 'Join Our Team' to 'Join the Team' and
'Slack' to 'Join the Slack' (also adds a permanent Slack invite link to
the footer instead of just the workspace URL).