wip: update appt creation flow
We removed the old parent-approval flow in place of:
1. The pupil (or parent) sends a request to a tutor.
2. We send an email to all the attendees of the appointment (i.e. the
   tutor) CC-ing the original creator of the appointment (i.e. the
   pupil).
3. The tutor and pupil then use that anonymous email relay system to
   communicate (e.g. to setup an appointment time and venue).

Note that there is still a big issue with this, as it assumes that
emails can be both uppercase and lowercase (because we're using a
`<uid>-<appt>@mail.tutorbook.org` format).

We have to revise the specs in #82 to address that lowercase issue.