fix(lib/model/account): revert to normal class
I made this an abstract class because it should never be created by
itself (rather, one should use the `Org` or `User` classes instead).

Making it abstract, however, prevented the `User` and `Org` classes from
ensuring that they don't override fields in the `Account` class. We have
to create an `Account()` instance to check what fields are already set
by that base class. Ideally, we'd be able to depend on the Typescript
`AccountInterface` to check those fields.