wip: migrate from 'react-intl' to 'next-translate'
Fixes #35 by replacing 'react-intl' with 'next-translate' (which is a
lot lighter weight and was made for use with Next.js).

Note that we still have an issue using the 'next-translate' <Link>
component and that we'll probably want to implement the "build step" as
a temporary workaround to enable automatic static optimization for our
dashboard pages.

We'll probably also add a redirect from '/en' to '/' and get rid of our
'/api/redirect' endpoint temporarily (while we don't support any other
locales). As soon as we get translations for other locales, however,
we'll add it back in.