feat: add org configurable required fields
Fixes #103. Orgs can now specify which profile fields are required (e.g.
setting the `phone` field as required).

If the `phone` field is set as required (i.e. is included in the org's
`profiles` array), the `/[org]/signup` page phone number field is
required and a phone number field is added to the `RequestDialog` shown
in the `/[org]/search` page.

Note: I still have to add server-side validation for all of these recent
changes.