fix(login): use custom hook to redirect to login
Create a custom hook `useLoggedIn` that redirects the user to the login
page using the specified `href` and `as` redirect parameters (i.e. the
page that the user should return to once they're signed in).

This also updates the redirect login in the `components/login` itself to
support dynamic paths using a `href` and `as` object.