fix(profile): patch security hole by locking email
Locks the user's email address after they login. This way, only verified
email addresses should ever be attached to an account (i.e. the user
must login with a verified provider to add an email address to TB).

Fixes #116. In the future, we'll want to allow the user to add or change
their email addresses but for now, we'll just lock them.