feat(search): add by-page pagination to search view
Fixes #110 using the same pagination component that's currently in use
by the "People" admin dashboard.

I figured that the `useSWRInfinite` wasn't worth the trouble to research
and implement. Plus, now we have a consistent UX (so someone familiar
with the "People" dashboard functionality can easily use the "Search"
view pagination and vice versa).