feat: allow orgs to configure available subjects
Fixes #133 by adding a 'subjects' property to the org data model and
then filtering by those subjects in the 'SubjectSelect' component.

Note that we still will have to eventually add i18n to those subjects
which will make this problem a lot harder to fix.

Also note that there currently isn't a way for org admins to customize
the subjects themselves (i.e. they have to have me do it for them).