fix(calendar): prevent subjects from disappearing
Previously, subjects would disappear because the SubjectSelect in the
EditDialog would be passed an empty options array. Thus, the
SubjectSelect wouldn't have any options to choose from and would try to
update the meeting's match's subjects to also be empty. This would
result in disappearing subjects and a React state update error (because
the SubjectSelect tried to update it's parent component while
rendering).

Now, we add the meeting's match's subjects to the meeting tutor/mentor
by default (before the actual person's data is fetched).