fix(user/display): place action buttons in bottom left
When admins click on a user profile, their mouse is usually closer to
the bottom left on the user's avatar than it is to where the action
buttons were earlier.

This change makes editing and vetting profiles faster (as you don't have
to move your mouse as much to click the "edit" button).