wip(time-select): users can only book a month into the future
Temporary front-end fix for #197. The best way to address this issue
would be to only send availability for the next 30 days and then
disable/enable the month pagination buttons based on the # of timeslots
fetched for that month (i.e. prefetch the prev/next month's timeslots
and then if there aren't any, then disable that pagination button).