ci(test.yml): remove parallelization for shared db
Don't run tests in parallel because they all use the same Supabase
database and thus the state of one test will affect another.

Once supabase/cli#45 is fixed, I should be able to use a local
PostgreSQL database and thus run as many tests in parallel as I want to.
This is definitely the preferred route to go because I will probably be
wanting to be testing multiple commits at the same time anyways.